+++ src/openrct2/rct2/S6Exporter.cpp
@@ -186,7 +186,7 @@ void S6Exporter::Export()
         }
     }
 
-    _s6.elapsed_months = gDateMonthsElapsed;
+    _s6.elapsed_months = static_cast<uint16_t>(gDateMonthsElapsed);
     _s6.current_day = gDateMonthTicks;
     _s6.scenario_ticks = gScenarioTicks;
